services/alignment/requirements.txt
psycopg2-binary==2.9.9
google-genai==0.6.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
nltk==3.8.1
pythainlp==5.1.2
python-crfsuite==0.9.10
python-dotenv==1.0.1
